adds duration, econ loss, and arrival rate scripts

+++ pandemic-loss-modeling/scripts/fit_distributions.py
@@ -97,8 +97,8 @@ if __name__ == "__main__":
         'dist_family': "Lognormal",
         'max_duration': float(df_all['duration'].max()), # With HIV/AIDS this is going to get us a massive number
         'params': {
-            'mu': np.log(params_all_risk[2]),
-            'sigma': params_all_risk[0]
+            'mu': float(np.log(params_all_risk[2])),
+            'sigma': float(params_all_risk[0])
         }
     }
 
@@ -106,8 +106,8 @@ if __name__ == "__main__":
         'dist_family': "Lognormal",
         'max_duration': float(df_resp['duration'].max()), # With HIV/AIDS this is going to get us a massive number
         'params': {
-            'mu': np.log(params_resp[2]),
-            'sigma': params_resp[0]
+            'mu': float(np.log(params_resp[2])),
+            'sigma': float(params_resp[0])
         }
     }
